Master

+++ TrainScheduling/src/main/java/com/railway/scheduling/DTO/Edge.java
@@ -12,7 +12,7 @@ import java.sql.Time;
 public class Edge {
     private Track track;
     private Float weight;
-    public Edge(Track track, int priority, int deviationPenalty, boolean isOnList, Time waitingTime){
+    public Edge(Track track, int priority, float deviationPenalty, boolean isOnList, Time waitingTime){
         this.track = track;
         this.weight = Formulas.weight(priority,deviationPenalty,isOnList,track.getTrackLength(),track.getMaximumSpeed(),waitingTime);
     }
